@@ -1,7 +1,11 @@
-package com.leonidas.helpdesk.api.security.controller;
+package com.leonidas.HelpDesk.api.security.controller;
 
-import javax.servlet.http.HttpServletRequest;
-
+import com.leonidas.HelpDesk.api.entity.User;
+import com.leonidas.HelpDesk.api.security.jwt.JwtAuthenticationRequest;
+import com.leonidas.HelpDesk.api.security.jwt.JwtTokenUtil;
+import com.leonidas.HelpDesk.api.security.model.CurrentUser;
+import com.leonidas.HelpDesk.api.service.UserService;
+import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.authentication.AuthenticationManager;
@@ -15,54 +19,47 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.leonidas.helpdesk.api.entity.User;
-import com.leonidas.helpdesk.api.security.jwt.JwtAuthenticationRequest;
-import com.leonidas.helpdesk.api.security.jwt.JwtTokenUtil;
-import com.leonidas.helpdesk.api.security.model.CurrentUser;
-import com.leonidas.helpdesk.api.service.UserService;
-
 @RestController
 @CrossOrigin(origins = "*")
 public class AuthenticationRestController {
 
-	@Autowired
-	private AuthenticationManager authenticationManager;
+    @Autowired
+    private AuthenticationManager authenticationManager;
 
-	@Autowired
-	private JwtTokenUtil jwtTokenUtil;
+    @Autowired
+    private JwtTokenUtil jwtTokenUtil;
 
-	@Autowired
-	private UserDetailsService userDetailsService;
+    @Autowired
+    private UserDetailsService userDetailsService;
 
-	@Autowired
-	private UserService userService;
+    @Autowired
+    private UserService userService;
 
-	@PostMapping(value = "/api/auth")
-	public ResponseEntity<?> createAuthenticationToken(@RequestBody JwtAuthenticationRequest authenticationRequest)
-			throws Exception {
-		final Authentication authentication = authenticationManager
-				.authenticate(new UsernamePasswordAuthenticationToken(authenticationRequest.getEmail(),
-						authenticationRequest.getPassword()));
-		SecurityContextHolder.getContext().setAuthentication(authentication);
-		final UserDetails userDetails = userDetailsService.loadUserByUsername(authenticationRequest.getEmail());
-		final String token = jwtTokenUtil.generateToken(userDetails);
-		final User user = userService.findByEmail(authenticationRequest.getEmail());
-		user.setPassword(null);
-		return ResponseEntity.ok(new CurrentUser(token, user));
-	}
+    @PostMapping(value = "/api/auth")
+    public ResponseEntity<?> createAuthenticationToken(@RequestBody JwtAuthenticationRequest authenticationRequest) throws Exception {
+        Authentication authentication = authenticationManager
+                .authenticate(new UsernamePasswordAuthenticationToken(authenticationRequest.getEmail(),
+                        authenticationRequest.getPassword()));
+        SecurityContextHolder.getContext().setAuthentication(authentication);
+        UserDetails userDetails = userDetailsService.loadUserByUsername(authenticationRequest.getEmail());
+        final String token = jwtTokenUtil.generateToken(userDetails);
+        final User user = userService.findByEmail(authenticationRequest.getEmail());
+        user.setPassword(null);
+        return ResponseEntity.ok(new CurrentUser(token, user));
+    }
 
-	@PostMapping(value = "/api/refresh")
-	public ResponseEntity<?> refreshAndGetAuthenticationToken(HttpServletRequest request) {
-		String token = request.getHeader("Authorization");
-		String username = jwtTokenUtil.getUsernameFromToken(token);
-		final User user = userService.findByEmail(username);
+    @PostMapping(value = "/api/refresh")
+    public ResponseEntity<?> refreshAndGetAuthenticationToken(HttpServletRequest request) {
+        String token = request.getHeader("Authorization");
+        String username = jwtTokenUtil.getUsernameFromToken(token);
+        final User user = userService.findByEmail(username);
 
-		if (jwtTokenUtil.canTokenBeRefreshed(token)) {
-			String refreshedToken = jwtTokenUtil.refreshToken(token);
-			return ResponseEntity.ok(new CurrentUser(refreshedToken, user));
-		} else {
-			return ResponseEntity.badRequest().body(null);
-		}
-	}
+        if (jwtTokenUtil.canTokenBeRefreshed(token)) {
+            String refreshedToken = jwtTokenUtil.refreshToken(token);
+            return ResponseEntity.ok(new CurrentUser(refreshedToken, user));
+        } else {
+            return ResponseEntity.badRequest().body(null);
+        }
+    }
 
 }
