@@ -1,7 +1,9 @@
-package com.leonidas.helpdesk.api.enums;
+package com.leonidas.HelpDesk.api.enums;
 
 public enum PriorityEnum {
 
-	High, Normal, Low;
+    HIGH,
+    NORMAL,
+    LOW
 
 }
