@@ -1,26 +1,31 @@
-package com.leonidas.helpdesk.api.enums;
+package com.leonidas.HelpDesk.api.enums;
 
 public enum StatusEnum {
 
-	New, Assigned, Resolved, Approved, Disapproved, Closed;
+    NEW,
+    ASSIGNED,
+    RESOLVED,
+    APPROVED,
+    DISAPPROVED,
+    CLOSED;
 
-	public static StatusEnum getStatus(String status) {
-		switch (status) {
-		case "New":
-			return New;
-		case "Assigned":
-			return Assigned;
-		case "Resolved":
-			return Resolved;
-		case "Approved":
-			return Approved;
-		case "Disapproved":
-			return Disapproved;
-		case "Closed":
-			return Closed;
-		default:
-			return New;
-		}
-	}
+    public static StatusEnum getStatus(String status){
+        switch (status){
+            case "New":
+                return NEW;
+            case "Assigned":
+                return ASSIGNED;
+            case "Resolved":
+                return RESOLVED;
+            case "Approved":
+                return APPROVED;
+            case "Disapproved":
+                return DISAPPROVED;
+            case "Closed":
+                return CLOSED;
+            default:
+                return null;
+        }
+    }
 
 }
