@@ -1,24 +1,22 @@
-package com.leonidas.helpdesk.api.repository;
+package com.leonidas.HelpDesk.api.repository;
 
+import com.leonidas.HelpDesk.api.entity.Ticket;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
-import com.leonidas.helpdesk.api.entity.Ticket;
-
 public interface TicketRepository extends MongoRepository<Ticket, String> {
 
-	Page<Ticket> findByUserIdOrderByDateDesc(Pageable pages, String userId);
+    Page<Ticket> findByUserIdOrderByDateDesc(Pageable pages, String userId);
 
-	Page<Ticket> findByTitleIgnoreCaseContainingAndStatusAndPriorityOrderByDateDesc(String title, String status,
-			String priority, Pageable pages);
+    Page<Ticket> findByTitleIgnoreCaseContainingAndStatusAndPriorityOrderByDateDesc(
+                   String title, String status, String priority, Pageable pages);
 
-	Page<Ticket> findByTitleIgnoreCaseContainingAndStatusAndPriorityAndUserIdOrderByDateDesc(String title,
-			String status, String priority, Pageable pages);
+    Page<Ticket> findByTitleIgnoreCaseContainingAndStatusAndPriorityAndUserIdOrderByDateDesc(
+            String title, String status, String priority, String userId, Pageable pages);
 
-	Page<Ticket> findByTitleIgnoreCaseContainingAndStatusAndPriorityAndAssignedUserIdOrderByDateDesc(String title,
-			String status, String priority, Pageable pages);
+    Page<Ticket> findByTitleIgnoreCaseContainingAndStatusAndPriorityAndAssigneeIdOrderByDateDesc(
+            String title, String status, String priority, String assigneeId, Pageable pages);
 
-	Page<Ticket> findByNumber(Integer number, Pageable pages);
-
+    Page<Ticket> findByNumber(Integer number, Pageable pages);
 }
