@@ -1,37 +1,20 @@
-package com.leonidas.helpdesk.api.security.jwt;
+package com.leonidas.HelpDesk.api.security.jwt;
 
+import lombok.*;
+
+import java.io.Serial;
 import java.io.Serializable;
 
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
 public class JwtAuthenticationRequest implements Serializable {
 
-	private static final long serialVersionUID = 1L;
+    @Serial
+    private static final long serialVersionUID = -438916409272195713L;
 
-	private String email;
-	private String password;
-
-	public JwtAuthenticationRequest() {
-		super();
-	}
-
-	public JwtAuthenticationRequest(String email, String password) {
-		this.setEmail(email);
-		this.setPassword(password);
-	}
-
-	public String getEmail() {
-		return email;
-	}
-
-	public void setEmail(String email) {
-		this.email = email;
-	}
-
-	public String getPassword() {
-		return password;
-	}
-
-	public void setPassword(String password) {
-		this.password = password;
-	}
+    private String email;
+    private String password;
 
 }
