@@ -1,113 +1,124 @@
-package com.leonidas.helpdesk.api.security.jwt;
-
-import java.io.Serializable;
-import java.util.Date;
-import java.util.HashMap;
-import java.util.Map;
-
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.stereotype.Component;
+package com.leonidas.HelpDesk.api.security.jwt;
 
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.SignatureAlgorithm;
+import io.jsonwebtoken.io.Decoders;
+import io.jsonwebtoken.security.Keys;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.stereotype.Component;
+
+import java.security.Key;
+import java.time.Instant;
+import java.time.LocalDateTime;
+import java.time.ZoneId;
+import java.time.format.DateTimeFormatter;
+import java.util.Date;
+import java.util.HashMap;
+import java.util.Map;
 
 @Component
-public class JwtTokenUtil implements Serializable {
+public class JwtTokenUtil {
 
-	private static final long serialVersionUID = 1L;
+    private static final DateTimeFormatter FORMATTER = DateTimeFormatter.ISO_LOCAL_DATE_TIME;
 
-	private final String CLAIM_KEY_USERNAME = "sub";
-	private final String CLAIM_KEY_CREATED = "created";
-	private final String CLAIM_KEY_EXPIRED = "exp";
+    static final String CLAIM_KEY_USERNAME = "sub";
+    static final String CLAIM_KEY_CREATED = "created";
+    static final String CLAIM_KEY_EXPIRED = "exp";
 
-	@Value("${jwt.secret}")
-	private String secret;
+    @Value("${security.jwt.secret-key}")
+    private String secret;
 
-	@Value("${jwt.expiration}")
-	private Long expiration;
+    @Value("${security.jwt.expiration-time}")
+    private Long expiration;
 
-	public String getUsernameFromToken(String token) {
-		String username;
-		try {
-			final Claims claims = getClaimsFromToken(token);
-			username = claims.getSubject();
-		} catch (Exception e) {
-			username = null;
-		}
-		return username;
-	}
+    public String getUsernameFromToken(String token){
+        String username;
+        try{
+            final Claims claims = getClaimsFromToken(token);
+            username = claims.getSubject();
+        } catch(Exception ex) {
+            username = null;
+        }
+        return username;
+    }
 
-	public Date getExpirationDateFromToken(String token) {
-		Date expiration;
-		try {
-			final Claims claims = getClaimsFromToken(token);
-			expiration = claims.getExpiration();
-		} catch (Exception e) {
-			expiration = null;
-		}
-		return expiration;
-	}
+    public Date getExpirationDateFromToken(String token) {
+        Date expiration;
+        try {
+            final Claims claims = null;
+            expiration = claims.getExpiration();
+        } catch(Exception ex) {
+            expiration = null;
+        }
+        return expiration;
+    }
 
-	private Claims getClaimsFromToken(String token) {
-		Claims claims;
-		try {
-			claims = Jwts.parser()//
-					.setSigningKey(secret)//
-					.parseClaimsJws(token)//
-					.getBody();
-		} catch (Exception e) {
-			claims = null;
-		}
-		return claims;
-	}
+    private Claims getClaimsFromToken(String token){
+        Claims claims;
+        try {
+            claims = Jwts.parser().setSigningKey(token).build().parseSignedClaims(token).getBody();
+        } catch(Exception ex){
+            claims = null;
+        }
+        return claims;
+    }
 
-	private Boolean isTokenExpired(String token) {
-		final Date expiration = getExpirationDateFromToken(token);
-		return expiration.before(new Date());
-	}
+    private Boolean isTokenExpired(String token){
+        final Date expiration = getExpirationDateFromToken(token);
+        return expiration.before(new Date());
+    }
 
-	public String generateToken(UserDetails userDetails) {
-		Map<String, Object> claims = new HashMap<>();
-		claims.put(CLAIM_KEY_USERNAME, userDetails.getUsername());
+    public String generateToken(UserDetails userDetails){
+        Map<String, Object> claims = new HashMap<>();
 
-		final Date createdDate = new Date();
-		claims.put(CLAIM_KEY_CREATED, createdDate);
+        final LocalDateTime createdDate = LocalDateTime.now();
+        claims.put(CLAIM_KEY_USERNAME, userDetails.getUsername());
+        claims.put(CLAIM_KEY_CREATED, createdDate.toString());
 
-		return doGenerateToken(claims);
-	}
+        return doGenerateToken(claims);
+    }
 
-	private String doGenerateToken(Map<String, Object> claims) {
-		final Date createdDate = (Date) claims.get(CLAIM_KEY_CREATED);
-		final Date expirationDate = new Date(createdDate.getTime() + expiration * 1000);
-		return Jwts.builder()//
-				.setClaims(claims)//
-				.setExpiration(expirationDate)//
-				.signWith(SignatureAlgorithm.HS512, secret)//
-				.compact();
-	}
+    private String doGenerateToken(Map<String, Object> claims) {
+        final String createdDateString = (String) claims.get(CLAIM_KEY_CREATED);
+        final LocalDateTime createdDate = LocalDateTime.parse(createdDateString, FORMATTER);
 
-	public Boolean canTokenBeRefreshed(String token) {
-		return (!isTokenExpired(token));
-	}
+        final Instant createdInstant = createdDate.atZone(ZoneId.systemDefault()).toInstant();
+        final Date expirationDate = Date.from(createdInstant.plusMillis(expiration * 1000));
 
-	public String refreshToken(String token) {
-		String refreshedToken;
-		try {
-			final Claims claims = getClaimsFromToken(token);
-			claims.put(CLAIM_KEY_CREATED, new Date());
-			refreshedToken = doGenerateToken(claims);
-		} catch (Exception e) {
-			refreshedToken = null;
-		}
-		return refreshedToken;
-	}
+        return Jwts.builder()
+                .setClaims(claims)
+                .setExpiration(expirationDate)
+                .signWith(getSignInKey(), SignatureAlgorithm.HS256)
+                .compact();
+    }
 
-	public Boolean validateToken(String token, UserDetails userDetails) {
-		JwtUser user = (JwtUser) userDetails;
-		final String username = getUsernameFromToken(token);
-		return (username.equals(user.getUsername()) && !isTokenExpired(token));
-	}
+    public Boolean canTokenBeRefreshed(String token){
+        return (!isTokenExpired(token));
+    }
+
+    public String refreshToken(String token){
+        String refreshedToken;
+        try {
+            final Claims claims = getClaimsFromToken(token);
+            claims.put(CLAIM_KEY_CREATED, new Date());
+            refreshedToken = doGenerateToken(claims);
+        } catch(Exception ex){
+            refreshedToken = null;
+        }
+        return refreshedToken;
+    }
+
+    public Boolean validateToken(String token, UserDetails userDetails) {
+        JwtUser user = (JwtUser) userDetails;
+        final String username = getUsernameFromToken(token);
+        return (username.equals(user.getUsername()) && !isTokenExpired(token));
+    }
+
+    private Key getSignInKey() {
+        byte[] keyBytes = Decoders.BASE64.decode(secret);
+        return Keys.hmacShaKeyFor(keyBytes);
+    }
 
 }
