@@ -1,70 +1,36 @@
-package com.leonidas.helpdesk.api.security.jwt;
+package com.leonidas.HelpDesk.api.security.jwt;
 
-import java.util.Collection;
-
+import lombok.AllArgsConstructor;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 
-import com.fasterxml.jackson.annotation.JsonIgnore;
+import java.io.Serial;
+import java.util.Collection;
 
+@AllArgsConstructor
 public class JwtUser implements UserDetails {
 
-	private static final long serialVersionUID = 1L;
+    @Serial
+    private static final long serialVersionUID = -9045512452548783761L;
 
-	private final String id;
-	private final String username;
-	private final String password;
-	private final Collection<? extends GrantedAuthority> authorities;
+    private final String id;
+    private final  String username;
+    private final String password;
+    private final Collection<? extends GrantedAuthority> authorities;
 
-	public JwtUser(String id, String username, String password, Collection<? extends GrantedAuthority> authorities) {
-		this.id = id;
-		this.username = username;
-		this.password = password;
-		this.authorities = authorities;
-	}
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities() {
+        return this.authorities;
+    }
 
-	@JsonIgnore
-	public String getId() {
-		return id;
-	}
+    @Override
+    public String getPassword() {
+        return this.password;
+    }
 
-	@Override
-	public String getUsername() {
-		return username;
-	}
-
-	@JsonIgnore
-	@Override
-	public boolean isAccountNonExpired() {
-		return true;
-	}
-
-	@JsonIgnore
-	@Override
-	public boolean isAccountNonLocked() {
-		return true;
-	}
-
-	@JsonIgnore
-	@Override
-	public boolean isCredentialsNonExpired() {
-		return true;
-	}
-
-	@JsonIgnore
-	@Override
-	public String getPassword() {
-		return password;
-	}
-
-	@Override
-	public Collection<? extends GrantedAuthority> getAuthorities() {
-		return authorities;
-	}
-
-	@Override
-	public boolean isEnabled() {
-		return true;
-	}
+    @Override
+    public String getUsername() {
+        return this.username;
+    }
 
 }
