@@ -1,19 +1,21 @@
-package com.leonidas.helpdesk.api.service;
+package com.leonidas.HelpDesk.api.service;
 
+import com.leonidas.HelpDesk.api.entity.User;
 import org.springframework.data.domain.Page;
+import org.springframework.stereotype.Component;
 
-import com.leonidas.helpdesk.api.entity.User;
+import java.util.Optional;
 
+@Component
 public interface UserService {
 
-	User findByEmail(String email);
+    User findByEmail(String email);
 
-	User createOrUpdate(User user);
+    User createOrUpdate(User user);
 
-	User findById(String id);
+    Optional<User> findById(String id);
 
-	void delete(String id);
+    void delete(String id);
 
-	Page<User> findAll(int page, int count);
-
+    Page<User> findAll(int page, int count);
 }
