@@ -1,45 +1,38 @@
-package com.leonidas.helpdesk.api.service.impl;
+package com.leonidas.HelpDesk.api.service.impl;
 
+import com.leonidas.HelpDesk.api.entity.User;
+import com.leonidas.HelpDesk.api.repository.UserRepository;
+import com.leonidas.HelpDesk.api.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
-import org.springframework.stereotype.Service;
+import org.springframework.stereotype.Component;
 
-import com.leonidas.helpdesk.api.entity.User;
-import com.leonidas.helpdesk.api.repository.UserRepository;
-import com.leonidas.helpdesk.api.service.UserService;
+import java.util.Optional;
 
-@Service
+@Component
 public class UserServiceImpl implements UserService {
 
-	@Autowired
-	private UserRepository userRepository;
+    @Autowired
+    private UserRepository userRepository;
 
-	@Override
-	public User findByEmail(String email) {
-		return this.userRepository.findByEmail(email);
-	}
+    public User findByEmail(String email) {
+        return this.userRepository.findByEmail(email);
+    }
 
-	@Override
-	public User createOrUpdate(User user) {
-		return this.userRepository.save(user);
-	}
+    public User createOrUpdate(User user) {
+        return this.userRepository.save(user);
+    }
 
-	@Override
-	public User findById(String id) {
-		return this.userRepository.findOne(id);
-	}
+    public Optional<User> findById(String id) {
+        return this.userRepository.findById(id);
+    }
 
-	@Override
-	public void delete(String id) {
-		this.userRepository.delete(id);
-	}
+    public void delete(String id) {
+        this.userRepository.deleteById(id);
+    }
 
-	@Override
-	public Page<User> findAll(int page, int count) {
-		Pageable pages = new PageRequest(page, count);
-		return this.userRepository.findAll(pages);
-	}
-
+    public Page<User> findAll(int page, int count) {
+        return this.userRepository.findAll(PageRequest.of(page, count));
+    }
 }
